Making code easier to read and use by separating the code into different files. As well as, creating a JSON file to simplify the for loops and create a level format that is specified in the JSON file. Using the JSON file, I will be able to draw things on the canvas.
