+++ public/js/main.js
@@ -1,43 +1,5 @@
-function loadImage(url){
-  return new Promise(resolve => {
-    const image = new Image();
-    image.addEventListener('load', () => {
-      resolve(image);
-    })
-    image.src = url;
-  })
-}
-
-class SpriteSheet {
-  constructor(image, width, height) {
-    this.image = image;
-    this.width = width;
-    this.height = height;
-    this.tiles = new Map();
-  }
-  define(name, x, y){
-    const buffer = document.createElement('canvas');
-    buffer.width = this.width;
-    buffer.height = this.height;
-    buffer
-          .getContext('2d')
-          .drawImage(
-                    this.image,
-                    x * this.width,
-                    y * this.height,
-                    this.width,
-                    this.height,
-                    0,
-                    0,
-                    this.width,
-                    this.height);
-    this.tiles.set(name, buffer);
-  }
-  draw(name, context, x, y) {
-    const buffer = this.tiles.get(name);
-    context.drawImage(buffer, x, y);
-  }
-}
+import SpriteSheet from './SpriteSheet.js';
+import {loadImage, loadLevel} from './loaders.js';
 
 const canvas = document.getElementById("screen");
 const context = canvas.getContext("2d");
@@ -48,12 +10,23 @@ loadImage('/img/tiles.png')
 .then(image => {
   const sprites = new SpriteSheet(image, 16, 16);
   sprites.define('ground', 0, 0);
-  sprites.draw('ground', context, 45, 62);
+  sprites.define('sky', 3, 23);
 
-  context.drawImage(image, 
-    0, 0, 
-    16, 16,
+  loadLevel('1-1')
+  .then(level => {
+    console.log(level);
+  });
 
-    32, 32, 
-    16, 16);
+  for (let x = 0; x < 25; ++x) {
+    for (let y = 0; y < 14; ++y) {
+      sprites.drawTile('sky', context, x, y);
+    }
+  }
+
+  for (let x = 0; x < 25; ++x) {
+    for (let y = 12; y < 14; ++y) {
+      sprites.drawTile('ground', context, x, y);
+    }
+  }
+
 })
