+++ public/levels/1-1.json
@@ -1,0 +1,13 @@
+{
+  "background": [
+        {
+          "tile": "sky",
+          "ranges": [
+              [
+                0, 25,
+                0, 14
+              ]
+          ]
+        }
+  ]
+}
